test: stage1 해결 - UserDao를 인터페이스로 변경
